+++ src/prototypes/entities.lua
@@ -3,62 +3,6 @@ local ICON_BASE = GRAPHICS_BASE .. "icons/"
 data:extend(
 {
   {
-    type = "assembling-machine",
-    name = "observatory",
-    icon = ICON_BASE .. "observatory.png",
-    flags = {"placeable-neutral", "placeable-player", "player-creation"},
-    max_health = 150,
-    corpse = "big-remnants",
-    dying_explosion = "medium-explosion",
-    collision_box = {{-1.2, -1.2}, {1.2, 1.2}},
-    selection_box = {{-1.5, -1.5}, {1.5, 1.5}},
-    minable = {mining_time = 1, result = "observatory"},
-    light = {intensity = 0.75, size = 8, color = {r = 1.0, g = 1.0, b = 1.0}},
-    animation =
-    {
-      layers =
-      {
-        {
-          filename = "__base__/graphics/entity/lab/lab.png",
-          width = 113,
-          height = 91,
-          frame_count = 33,
-          line_length = 11,
-          animation_speed = 1 / 3,
-          shift = {0.2, 0.15}
-          -- hr_version = { scale = 0.5 } -- see assembling machine
-        }
-      }
-    },
-    crafting_speed = 1,
-    fixed_recipe = "observatory-scan-for-sites",
-    working_sound =
-    {
-      sound =
-      {
-        filename = "__base__/sound/lab.ogg", -- TODO: Custom sound
-        volume = 0.7
-      },
-      apparent_volume = 1
-    },
-    energy_source =
-    {
-      type = "electric",
-      usage_priority = "secondary-input"
-    },
-    energy_usage = "60kW",
-    crafting_categories = {"astronomy"},
-    vehicle_impact_sound =  { filename = "__base__/sound/car-metal-impact.ogg", volume = 0.65 },
-    ingredient_count = 1,
-    module_specification =
-    {
-      module_slots = 2
-    },
-    -- Note: No productivity, since inputs are virtual it's meaningless
-    -- TODO: *However*, maybe an actual input such as space science pack could actually make sense?
-    allowed_effects = {"consumption", "speed", "pollution"}
-  },
-  {
     type = "electric-energy-interface",
     name = "medium-portal",
     icon = ICON_BASE .. "medium-portal.png",
@@ -170,7 +114,7 @@ data:extend(
     --max_health = 150,
     --corpse = "medium-remnants",
     collision_box = {{0, 0}, {0, 0}},
-    --selection_box = {{-0.2, -0.2}, {0.2, 0.2}},
+    selection_box = {{-0.2, -0.2}, {0.2, 0.2}},
     -- Can't be selected because it's masked by the chest (for unknown reasons but still), however
     -- the selection_box is used to draw the power connection blue square so it's good to have one.
     --selection_box = {{-0.5, -0.5}, {0.5, 0.5}},